fix(#5919): widen numeric narrowing at LIMIT/SKIP, map2json, and DOUBLE MIN/MAX validation (#5921)
* fix(#5919): widen numeric narrowing at LIMIT/SKIP, map2json, and DOUBLE MIN/MAX validation

Follow-up to #5900/#5905/#5906: three more sites narrowed a wider numeric
without a range check.

- Limit/Skip.getValue() used intValue() on a value that can be a Long
  outside int range (e.g. LIMIT 2147483648), wrapping to a negative int.
  For LIMIT that made LimitExecutionStep return 0 rows instead of the
  full result; for SKIP it made SkipExecutionStep skip nothing instead of
  everything. Both now saturate to Integer.MAX_VALUE/MIN_VALUE instead of
  wrapping, without touching the -1 "unlimited" sentinel.
- JsonSerializer.map2json() classified finiteness via floatValue(), so a
  finite double above Float.MAX_VALUE narrowed to Infinity and was
  silently dropped from the serialized map. Now uses doubleValue().
- DocumentValidator's DOUBLE MIN/MAX checks compared fieldValue.floatValue()
  against a double bound, so a value beyond +/-Float.MAX_VALUE could be
  falsely rejected even when within bounds. Now compares doubleValue().

* refactor(#5919): share LIMIT/SKIP int-saturation via NumberUtils, cover BigInteger/BigDecimal bind params

Addresses PR #5921 code review feedback:
- Limit.saturateToInt() and Skip.saturateToInt() were byte-for-byte
  duplicates; moved to NumberUtils.saturateToInt(), which the project
  already uses for small numeric-parsing helpers.
- Number.longValue() on a BigInteger/BigDecimal far outside the Long
  range is documented as lossy in an unspecified way (it can even flip
  sign), so an out-of-range BigInteger/BigDecimal LIMIT/SKIP bind param
  could have produced a wrong non-saturated int instead of clamping.
  saturateToInt() now compares BigInteger/BigDecimal by magnitude
  directly instead of going through longValue() first.

(cherry picked from commit 86835bcd3159de1e4419af265d795e5187eb4fe9)

+++ engine/src/main/java/com/arcadedb/database/DocumentValidator.java
@@ -121,7 +121,7 @@ public class DocumentValidator {
     }
     case DOUBLE -> {
       final double maxAsDouble = Double.parseDouble(max);
-      if (((Number) fieldValue).floatValue() > maxAsDouble)
+      if (((Number) fieldValue).doubleValue() > maxAsDouble)
         throwValidationException(document.getType(), p, "value " + fieldValue + " is greater than " + max);
     }
     case DECIMAL -> {
@@ -200,7 +200,7 @@ public class DocumentValidator {
       }
       case DOUBLE -> {
         final double minAsDouble = Double.parseDouble(min);
-        if (((Number) fieldValue).floatValue() < minAsDouble)
+        if (((Number) fieldValue).doubleValue() < minAsDouble)
           yield new ValidationResult(true, "value " + fieldValue + " is less than " + min);
         yield new ValidationResult(false, null);
       }
